+++ developers/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import DeveloperList
+
+urlpatterns = [
+    path('', DeveloperList.as_view(), name='devs'),
+]