+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+
+- [codice-14] improve home dashboard
+
 ## [0.1.0] - 2020-03-16
 
 ### Added